+++ tests/integration/test___init__.py
@@ -13,18 +13,8 @@ from urllib.parse import urlparse
 
 import pytest
 from git.repo import Repo
-from github.PullRequest import PullRequest
 
-from src import (
-    GETTING_STARTED,
-    exceptions,
-    index,
-    metadata,
-    migration,
-    pull_request,
-    reconcile,
-    run,
-)
+from src import GETTING_STARTED, exceptions, index, metadata, reconcile, run
 from src.discourse import Discourse
 
 from .. import factories
@@ -377,113 +367,3 @@ async def test_run(
     assert_substrings_in_string(chain(urls, ("Update", "'success'")), caplog.text)
     index_topic = discourse_api.retrieve_topic(url=index_url)
     assert index_content not in index_topic
-
-
-@pytest.mark.asyncio
-@pytest.mark.usefixtures("patch_create_repository_client")
-async def test_run_migrate(
-    discourse_hostname: str,
-    discourse_api: Discourse,
-    caplog: pytest.LogCaptureFixture,
-    repository: tuple[Repo, Path],
-    upstream_repository: tuple[Repo, Path],
-    mock_pull_request: PullRequest,
-):
-    """
-    arrange: given running discourse server
-    act: when run is called with:
-        1. with no docs dir and no custom branchname provided
-        2. with no docs dir and custom branchname provided
-        3. with no changes applied after migration
-    assert: then:
-        1. the documentation files are pushed to default branch
-        2. the documentation files are pushed to custom branch
-        3. no operations are taken place
-    """
-    document_name = "migration name 1"
-    discourse_prefix = f"http://{discourse_hostname}"
-    (repo, repo_path) = repository
-    (upstream_repo, upstream_repo_path) = upstream_repository
-    content_page_1 = factories.ContentPageFactory()
-    content_page_1_url = discourse_api.create_topic(
-        title=content_page_1.title,
-        content=content_page_1.content,
-    ).removeprefix(discourse_prefix)
-    content_page_2 = factories.ContentPageFactory()
-    content_page_2_url = discourse_api.create_topic(
-        title=content_page_2.title,
-        content=content_page_2.content,
-    ).removeprefix(discourse_prefix)
-    content_page_3 = factories.ContentPageFactory()
-    content_page_3_url = discourse_api.create_topic(
-        title=content_page_3.title,
-        content=content_page_3.content,
-    ).removeprefix(discourse_prefix)
-    content_page_4 = factories.ContentPageFactory()
-    content_page_4_url = discourse_api.create_topic(
-        title=content_page_4.title,
-        content=content_page_4.content,
-    ).removeprefix(discourse_prefix)
-    index_page_content = f"""Testing index page.
-
-    Testing index page content.
-
-    # Navigation
-
-    | Level | Path | Navlink |
-    | -- | -- | -- |
-    | 1 | group-1 | [Group 1]() |
-    | 1 | group-2 | [Group 2]() |
-    | 2 | group-2-content-1 | [Content Link 1]({content_page_1_url}) |
-    | 2 | group-2-content-2 | [Content Link 2]({content_page_2_url}) |
-    | 1 | group-3 | [Group 3]() |
-    | 2 | group-3-group-4 | [Group 4]() |
-    | 3 | group-3-group-4-content-3 | [Content Link 3]({content_page_3_url}) |
-    | 2 | group-3-content-4 | [Content Link 4]({content_page_4_url}) |
-    | 1 | group-5 | [Group 5]() |"""
-    index_url = discourse_api.create_topic(
-        title=f"{document_name.replace('-', ' ').title()} Documentation Overview",
-        content=index_page_content,
-    )
-
-    # 1. with no docs dir and a metadata.yaml with docs key
-    caplog.clear()
-    create_metadata_yaml(
-        content=f"{metadata.METADATA_NAME_KEY}: name 1\n{metadata.METADATA_DOCS_KEY}: {index_url}",
-        path=repo_path,
-    )
-
-    urls_with_actions = run(
-        base_path=repo_path,
-        discourse=discourse_api,
-        user_inputs=factories.UserInputFactory(),
-    )
-
-    upstream_repo.git.checkout(pull_request.DEFAULT_BRANCH_NAME)
-    upstream_doc_dir = upstream_repo_path / index.DOCUMENTATION_FOLDER_NAME
-    assert tuple(urls_with_actions) == (mock_pull_request.html_url,)
-    assert ((group_1_path := upstream_doc_dir / "group-1")).is_dir()
-    assert (group_1_path / migration.GITKEEP_FILENAME).is_file()
-    assert ((group_2_path := upstream_doc_dir / "group-2")).is_dir()
-    assert (group_2_path / "content-1.md").read_text(encoding="utf-8") == content_page_1.content
-    assert (group_2_path / "content-2.md").read_text(encoding="utf-8") == content_page_2.content
-    assert ((group_3_path := upstream_doc_dir / "group-3")).is_dir()
-    assert ((group_4_path := group_3_path / "group-4")).is_dir()
-    assert (group_4_path / "content-3.md").read_text(encoding="utf-8") == content_page_3.content
-    assert (group_3_path / "content-4.md").read_text(encoding="utf-8") == content_page_4.content
-    assert (group_5_path := upstream_doc_dir / "group-5").is_dir()
-    assert group_5_path.is_dir()
-
-    # 2. with no changes applied after migration
-    caplog.clear()
-    repo.git.checkout(pull_request.DEFAULT_BRANCH_NAME)
-
-    urls_with_actions = run(
-        base_path=repo_path,
-        discourse=discourse_api,
-        user_inputs=factories.UserInputFactory(),
-    )
-
-    assert_substrings_in_string(
-        chain(urls_with_actions, ("Noop", "Noop", "Noop", "'success'")), caplog.text
-    )
